refactor : 로또 묶음 비교 책임 위임
