+++ src/main/java/model/vo/Lottery.java
@@ -1,0 +1,55 @@
+package model.vo;
+
+import model.Winning;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+
+public class Lottery {
+
+    private final List<Lotto> lottos;
+
+    public Lottery(List<Lotto> lottos) {
+        this.lottos = lottos;
+    }
+
+    public Map<Winning, Integer> drawLottery(Lotto winningNumbers, LottoNumber bonusNumber) {
+        Map<Winning, Integer> winningResult = Winning.initialSetting();
+        List<Integer> matchWithWinningNumbersResult = new ArrayList<>();
+        List<Boolean> matchWithBonusNumberResult = new ArrayList<>();
+
+        for (Lotto lotto : this.lottos) {
+            matchWithWinningNumbersResult.add(lotto.drawLottoWithWinningNumbers(winningNumbers));
+            matchWithBonusNumberResult.add(lotto.drawLottoWithBonusNumber(bonusNumber));
+        }
+
+        for (int index = 0; index < matchWithWinningNumbersResult.size(); index++) {
+            winningResult.put(addResult(matchWithWinningNumbersResult.get(index), matchWithBonusNumberResult.get(index)), 1);
+        }
+
+        return winningResult;
+    }
+
+    private Winning addResult(int matchCountOfWinningNumbers, boolean matchCountOfBonusNumber) {
+        return Winning.matchWinning(matchCountOfWinningNumbers, matchCountOfBonusNumber);
+    }
+
+    public List<Lotto> getLottery() {
+        return lottos;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Lottery lottery = (Lottery) o;
+        return Objects.equals(lottos, lottery.lottos);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(lottos);
+    }
+}
